==> 03. Scripts/01. F1 Data Cleaning.ipynb <==
# Formula 1 Data Cleaning

## Cleaning
### 01. Import Libraries

import pandas as pd
import numpy as np
import os

### 02. Data

#starting with IDs for Drivers, Constructors, and Circuits
dv = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\drivers.csv')
cn = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\constructors.csv')
crc = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\circuits.csv')

dv.head()

dv.value_counts

#dropping url columns
dv = dv.drop('url', axis = 1)
cn = cn.drop('url', axis = 1)
crc = crc.drop('url', axis = 1)
dv.head(1)

#removing duplicates
dv = dv.drop_duplicates()
cn = cn.drop_duplicates()
crc = crc.drop_duplicates()

#export cleaned ID sets
dv.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\driverID.pkl')
cn.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\constructorID.pkl')
crc.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\circuitID.pkl')

#importing seasons and races
sns = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\seasons.csv')
rcs = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\races.csv')

#dropping url columns
sns = sns.drop('url', axis = 1)
rcs = rcs.drop('url', axis = 1)

#exporting seasons
sns.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\seasons.pkl')

rcs.head()

rcs.dtypes

#dropping free practice time and date columns
rcs = rcs.drop(['fp1_time','fp1_date','fp2_date','fp2_time','fp3_time','fp3_date'], axis = 1)
#these columns do not contribute to driver or vehicle statistics 
rcs.head()

#exporting race
rcs.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\races.pkl')

#importing qualifying, sprint results, and status
qul = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\qualifying.csv')
spr_res = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\sprint_results.csv')
#note: a sprint is a short race (100km) and awards points towards the drivers championship
status = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\status.csv')

status.value_counts

#dropping duplicates
status = status.drop_duplicates()

qul.shape

spr_res.shape

#dropping duplicates
qul = qul.drop_duplicates()
spr_res = spr_res.drop_duplicates()

qul.shape

spr_res.shape

#exporting qualifying, sprint results, and status
qul.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\qualifying.pkl')
spr_res.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\sprint_results')
status.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\status.pkl')

#importing results, pit stops, and lap times
res = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\results.csv')
pit = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\pit_stops.csv')
lapt = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\lap_times.csv')

res = res.drop_duplicates()
pit = pit.drop_duplicates()
lapt = lapt.drop_duplicates()

res.head()

res.shape

pit.head()

lapt.head()

#exporting results, pit stops, and lap times
res.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\race_results.pkl')
pit.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\pit_stops.pkl')
lapt.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\lap_times.pkl')

#importing driver standings, constructor standings, and constructor results
dstand = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\driver_standings.csv')
cstand = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\constructor_standings.csv')
cres = pd.read_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Org Data\constructor_results.csv')

#removing duplicates
dstand = dstand.drop_duplicates()
cstand = cstand.drop_duplicates()
cres = cres.drop_duplicates()

#exporting driver standings, constructor standings, and constructor results.
dstand.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\driver_standings.pkl')
cstand.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\constructor_standings.pkl')
cres.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\constructor_results.pkl')
