==> 03. Scripts/02.1 F1 Project Data Merge.ipynb <==
# F1 Project Data Merger
## 01. Import Libraries

import pandas as pd
import numpy as np
import os

## 02. Import Data Sets

drid = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\driverID.pkl')
cnid = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\constructorID.pkl')
crcid = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\circuitID.pkl')
rcs = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\races.pkl')
qul = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\qualifying.pkl')
pit = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\pit_stops.pkl')
res = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\race_results.pkl')
lapt = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\lap_times.pkl')

drid.head()

## 03. Data Merger

#merging driverID on results
dr_merge = drid.merge(res, on = ['driverId','number'])
#removing unwanted column
dr_merge = dr_merge.drop('dob', axis = 1)
dr_merge

dr_merge.columns

#renaming column to resolve potential conflict
cnid = cnid.rename(columns = {'nationality':'con_nat'})
cnid.head()

#exporting change
cnid.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\constructorID.pkl')

#merging constructorID to dr_merge
dr_merge = dr_merge.merge(cnid, on = 'constructorId')
dr_merge

#creating merged set for circuits and races
rcs_merge = rcs.merge(crcid, on = ['circuitId'])
rcs_merge

#renaming columns
rcs_merge = rcs_merge.rename(columns = {'name_x':'gp_name', 'name_y':'circ_name'})

rcs_merge

dr_merge.columns

#merging dr_merge and rcs_merge
formula_full = dr_merge.merge(rcs_merge, on = 'raceId')
formula_full.head()

formula_full.shape

formula_full.columns

#exporting backup
formula_full.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\fullsetbackup.pkl')
