==> 03. Scripts/02.2 Lap Time Merge.ipynb <==
# F1 Project Lap Times
## 01. Import Libraries

import pandas as pd
import numpy as np
import os

## 02. Import Data

lapt = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\lap_times.pkl')
drid = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\driverID.pkl')
cnid = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\constructorID.pkl')
crcid = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\circuitID.pkl')
rcs = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\races.pkl')

lapt

## 03. Merger

drid.columns

 drid = drid.drop('dob', axis = 1)

#adding driver information
lap_merge = drid.merge(lapt, on = 'driverId')
lap_merge

#importing additional data frame
rcsres = pd.read_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\race_results.pkl')

#merging race results with constructor id 
res_merge = rcsres.merge(cnid, on = 'constructorId')
res_merge

res_merge.columns

rcs.columns

rcs_cir = rcs.merge(crcid, on = 'circuitId')
rcs_cir

rcs_cir = rcs_cir.drop('time', axis = 1)
rcs_cir = rcs_cir.rename(columns = {'name_x':'gp_name', 'name_y':'circuit'})
rcs_cir.head()

rcs_cir.columns

res_merge.columns

res_merge2 = res_merge.merge(rcs_cir, on = 'raceId')
res_merge2

res_merge2.columns

lap_merge.columns

lap_merge = lap_merge.rename(columns = {'lap':'lap_num', 'time':'lap_time', 'milliseconds':'lap_milli', 'position':'lap_pos'})

#merging lap times with race results
df_full = res_merge.merge(lap_merge, on = ['driverId','number','raceId'])

df_full

df_full.columns

## 04. Export

df_full.to_pickle(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\race_lap_full.pkl')
df_full.to_csv(r'C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\full_check.csv')
#export to csv for manual check.
