==> 03. Scripts/02.3 Driver Specific Analysis.ipynb <==
# Driver Specific Analysis
## 01. Import Libraries

import pandas as pd
import numpy as np
import os
import matplotlib
import matplotlib.pyplot as plt
import seaborn as sns

## 02. Data

res = pd.read_csv(r"C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\res_sub.csv")

res.head(1)

res.columns

dr = pd.read_pickle(r"C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\driverID.pkl")
cn = pd.read_pickle(r"C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\constructorID.pkl")

dr.head(1)

dr = dr.drop(['number','dob','nationality','forename','surname'], axis = 1)
dr.head(1)

cn.head(1)

cn = cn.drop('con_nat', axis = 1)

merge = cn.merge(res, on = 'constructorId')
merge.head(1)

merge = dr.merge(merge, on = 'driverId')
merge.head(1)

merge.to_csv(r"C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\F1Merge.csv")
