==> 03. Scripts/07. Conversion for Tableau.ipynb <==
# Converting Data to CSV for Tableau
## 01. Import Libraries

import pandas as pd
import numpy as np
import os

## 02. Data

df = pd.read_pickle(r"C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\race_lap_full.pkl")
df.head(10)

df.shape

## 03. Export to CSV

df.to_csv(r"C:\Users\jmanc\OneDrive\Desktop\F1 Project\Data\Prepped Data\race_lap_full.csv")
